Switch to an external file for upload request readiness and parameters. (#3)
The main loop now just has to watch for 'request-upload-*' files and parse the
JSON within to obtain the request details. In effect, these files are the HTTP
request body, with the file name indicating the endpoint - uploads here, but
this approach generalizes to other actions like permission setting.

Each request specifies the directory containing the files to be uploaded. We
"authenticate" by checking that the same user owns the request and files.
This eliminates the need to add watchers in the new upload subdirectory,
along with the need to explicitly skip the request file during transfer.