Distinguish between uploader entries in the request vs ..permissions.
In the former, the user ID might not be present and thus must be a
nil-able pointer, while in the latter, the ID is guaranteed to be
present and does not need to be checked. Thus, the use of a pointer
for the user ID should be strictly limited to request parsing.