Validate permissions of the responses subdirectory.
This ensures that (i) the calling process owns the subdirectory and that
(ii) other users can't just randomly write stuff into it. This is
important as clients will be looking for files in this subdirectory that
have the same name as their request file; if a malicious user can
write into this subdirectory, they could just make up fake responses.