Switched from flock to a single mutex for locking projects.
This protects against multiple goroutines attempting to acquire the lock
from different goroutines, which is not possible with flock (as all
goroutines comes from the same process so any lock that should contend
just becomes a no-op). It also eliminates the requirement for a
lock-capable filesystem; the cost is that we are no longer thread-safe
for multiple gobbler processes, but that shouldn't really happen anyway.