Limit creation of new projects to administrators, a la gypsum. (#10)
This aligns with the gypsum philosophy and gives administrators more control by
preventing users from just creating loads of projects; any projects that are
created must have some owners that are held accountable for usage. The change
also simplifies matters by separating the upload from the setup of permissions.

We also eliminated the idea of incremental projects. Given that an
administrator has to create the project anyway, there isn't much benefit to
automating the name creation. While we're here, we might as well get rid of the
incremental versions and simplify the code.