Replace fsnotify with a HTTP request to trigger actions. (#11)
Most shared filesystems are networked in some manner and don't share events
across nodes; this means that fsnotify doesn't actually work. Polling is too
taxing so instead we ask the client to ping an API to indicate that a request
body has been written to the staging directory and is ready for execution. 

We don't ask the client to provide the request details in the API as it's
unauthenticated. We're still relying on the Unix file owner to tell us who is
making the request; the ping just tells us that the file is ready, which
happily eliminates the need for the retry loop. The output is also returned
as a HTTP response, which eliminates the need for the responses directory.

Some extra work is involved in making sure that the correct HTTP status codes
are reported for the different errors. We also mandate go >= 1.22.1 now.