Improved global write permissions to avoid updates from other users.
With the * wildcard, any user could update any other user's assets, which is
not desirable. Instead, we introduce a global_write=true option in the
permissions, which instructs gobbler to mutate the permissions by adding a
new entry to the uploaders array when a new asset is created. This allows
any user to create a new asset but only allows that same user to update it
(unless the permissions are explicitly modified to allow multiple users).