Use asset-level permissions to implement global_write. (#19)
This introduces an asset-level ..permissions file containing uploaders that are
appended to the project-level permissions during an upload request. It reduces
the size of the project-level ..permissions when global_write is enabled, as we
don't accumulate everyone's asset permissions into the same file. This should
improve scalability for projects that have lots of different uploaders.