Added option to forcibly delete versions/assets or reject probational status.
If the manifest/summary files aren't available, this probably indicates that
there's something wrong with the version directory, in which case it should be
deleted. However, the deletion/rejection endpoints rely on these files to do
various things (e.g., updating the project usage).

Setting force=true will skip the additional checks if the manifest/summary
files aren't present to ensure that the version can be forcibly deleted.
The default is still force=false so that the caller can be notified of a
problem before possibly repeating the request with force=true.