Implement a whitelist for links outside of the registry. (#20)
Any symlinks in the registry that target files inside the whitelisted
directories are treated as real files for purposes of upload/reindexing. This
allows us to easily refer to big files in immutable external archives without
forcing them to be directly stored and managed by the gobbler. 

Note that this change also involved consolidating the reindexing and transfer
code back into a single internal function, to avoid duplicated logic.