Update project usage statistics when rerouting creates a copy.
We also return the increase in project usage when a copy is created, which
gives admins a way to judge whether it is worth deleting a directory to free up
disk space if that saving would be offset by a increased number of copies.

Also lock the directory for each project when it's being rerouted,
so that other project-level modifications don't interfere.