Set a registry-wide exclusive lock when deleting projects.
We also use a shared lock for all other endpoints that access individual
projects. This ensures that the deletion of a project doesn't interfere with
other endpoints that assume that the directories are still there.

We extend the use of the exclusive registry-wide lock to the rerouting
endpoint, as this could affect an unknown number of projects; and to the
project addition endpoint, to avoid conflicts between addition/deletion.