Added consume= and ignore_dot= options to the upload endpoint.
The consume= option allows the Gobbler to attempt to move files from the
staging directory into the registry. This is useful in the case where the two
directories live on the same device, where a move would save time and storage.
If it fails (e.g., as the two directories live on different devices), the
Gobbler will silently fall back to a copy.

The ignore_dot= option instructs the Gobbler to ignore dotfiles during
transfer. This is helpful for clients to protect against naive users who try to
upload directories containing .git or other crap. It effectively restores the
previous "ignore all dotfiles" behavior that used to be the default.