Don't transfer symlinks to non-whitelisted directories during upload.
This closes a potential security hole whereby a user creates a link to a
non-public file that is accessible by the service account running the Gobbler.
For the user, the link is correctly invalid, but the service account would
happily copy the target into the registry, where it becomes visible to all.