Clean all user-supplied paths for consistency.
This avoids problems when the registry path is extracted from /info and used
for string matching to the cleaned file paths in other systems, e.g., SewerRat.

The use of cleaned paths in the link whitelist also requires proper filepath
comparisons in the whitelist checks, which is more robust than string matching.