More fine-grained locks to improve parallelism across endpoints. (#37)
We force every function to explicitly acquire the registry lock, to make it
more obvious that directory locks are acquired hierarchically. We also acquire
locks on the asset directory where relevant, to improve parallelization with
other functions that are working on different assets in the same project.

We modify the lock calls so that repeated unlocks are a no-op, allowing us to
easily unlock resources without waiting for the defer in a anonymous function.
This allows us to easily switch the upload handler's project-level lock from
exclusive to shared to enable parallel transfers after asset initialization.

We introduce the concept of a usage lock for modifying the project-level usage
file. This allows several handlers to operate in parallel with shared project
locks, only requiring synchronization when they edit the project's usage file.
Also rolled out the usage editing code into a common function for convenience.

We reacquire the project-level shared lock when processing each asset in the
probation purging function, to provide some opportunities for contending
handlers to acquire the lock while the purge is in progress. This prevents
other handlers from being blocked for a long time during a project-level purge.

We clean up some of the directory existence checks to give an appropriate error
if the directory could not be stat'd.