Add concurrent processing to file-intensive endpoints. (#38)
This applies to the loops across files in the upload and reindexing endpoints,
plus the loops across versions in the rerouting endpoint and probational
purges. We throttle the concurrency to ensure that we don't put too much stress
on the filesystem, e.g., by exceeding the maximum number of open filehandles.