Refactor the directory walk for more versatile handling of symlinks. (#40)
This allows us to switch between creation of a new symlink, replacement of an
existing symlink, or a no-op, depending on what we want to do.

The idea is to prepare for the easy creation of a new integrity-check endpoint
that constructs a new manifest from the existing directory contents for
comparison to the actual manifest, without modifying any aspects of the
directory (e.g., no creation of new symlinks).

We also flesh out tests for preservation of ancestral links after reindexing.